// server/src/utils/searchers/kinguin/types.d.ts
// Generated by https://quicktype.io

export interface KinguinResult {
    _embedded: Embedded;
    _links: KinguinResultLinks;
    page: Page;
}

export interface Embedded {
    products: Product[];
    info: Info;
}

export interface Info {
    taxRate: number;
    currencyRate: number;
}

export interface Product {
    id: string;
    vendor: Vendor;
    name: string;
    externalId: string;
    imageUrl: string;
    hiImageUrl: string;
    attributes: Attributes;
    price: Price;
    active: boolean;
    delivery: number[];
    version: number;
    stockChangeVersion: number;
    categories: number[];
    visible: boolean;
    checkoutTypes: CheckoutType[];
    offerId: string;
    offerChecksum?: string;
    wholesale: Wholesale;
    names: any;
    tags: any[];
    isFinal: boolean;
    breadcrumbs: Breadcrumb[];
    isDeleted: boolean;
    createDate: string;
    _links: ProductLinks;
}

export interface ProductLinks {
    self: Self;
    product: Self;
}

export interface Self {
    href: string;
}

export interface Attributes {
    platform: number;
    genres: number[];
    region: Region;
    parentCategoryId: number;
    languages: string[];
    systems: System[];
    types: number[];
    restrictions?: Restrictions;
    bundle: boolean;
    keywords: string;
    urlKey: string;
    releaseDate: string;
    displayMode: DisplayMode;
    isKinguinWebsite: boolean;
    marketingProductType: MarketingProductType;
    metascore?: number;
    igdbTags?: string[];
}

export enum DisplayMode {
    Products = "products",
}

export enum MarketingProductType {
    Preorder = "PREORDER",
}

export interface Region {
    id: number;
    excludedCountries: string[];
}

export interface Restrictions {
    countries: string[];
    minAge: null;
    store: null;
}

export enum System {
    Windows = "Windows",
}

export interface Breadcrumb {
    id: number;
    productId: string;
    name: string;
    slug: string;
}

export enum CheckoutType {
    BuyNow = "buyNow",
}

export interface Price {
    lowestOffer: number;
    discount?: number;
    market: number;
    calculated: number;
}

export enum Vendor {
    Kinguin = "kinguin",
}

export interface Wholesale {
    enabled: boolean;
    lowestPrice: number;
}

export interface KinguinResultLinks {
    self: Self;
}

export interface Page {
    size: number;
    totalElements: number;
    totalPages: number;
    number: number;
}